Add initial clippy checks

+++ core/src/types/server_player.rs
@@ -120,6 +120,6 @@ impl ServerPlayer {
     }
 
     pub fn is_sane_player(player_index: usize) -> bool {
-        player_index < crate::constants::MAXPLAYER as usize
+        player_index < crate::constants::MAXPLAYER
     }
 }
