+++ server/src/area.rs
@@ -1020,7 +1020,7 @@ const AREAS: &[Area] = &[
 ];
 
 pub fn is_in_pentagram_quest(cn: usize) -> bool {
-    if cn < 1 || cn >= crate::core::constants::MAXCHARS as usize {
+    if cn < 1 || cn >= crate::core::constants::MAXCHARS {
         return false;
     }
 
@@ -1041,7 +1041,7 @@ pub fn is_in_pentagram_quest(cn: usize) -> bool {
 // Unused in original implementation as well
 #[allow(dead_code)]
 pub fn get_area(cn: usize, verbose: bool) -> String {
-    if cn < 1 || cn >= crate::core::constants::MAXCHARS as usize {
+    if cn < 1 || cn >= crate::core::constants::MAXCHARS {
         return String::new();
     }
